updated tof distance to 16bit

+++ include/tof_sensor.h
@@ -4,6 +4,6 @@
 #include <Arduino.h>
 
 void initToFSensor();
-void readToF(uint8_t &d1, uint8_t &d2, uint8_t &d3, String &l, String &r, String &f);
+void readToF(uint16_t &d1, uint16_t &d2, uint16_t &d3, String &l, String &r, String &f);
 
 #endif
